Refactored code for better separation
